// packages/@higuma/fig-builder/src/types/spec-types.ts
export type BaseNodeSpec = {
  readonly name?: string;
  readonly x: number;
  readonly y: number;
  readonly width: number;
  readonly height: number;
  readonly rotation?: number;
  readonly fills?: readonly FigPaint[];
  readonly strokes?: readonly FigPaint[];
  readonly strokeWeight?: number;
  readonly effects?: readonly FigEffect[];
  readonly opacity?: number;
  readonly visible?: boolean;
};
// ...
export type FrameNodeSpec = BaseNodeSpec & {
  readonly type: "FRAME";
  readonly clipsContent?: boolean;
  readonly autoLayout?: AutoLayoutProps;
  readonly backgroundColor?: FigColor;
};